+++ src/lib/messages/utils.js
@@ -1,5 +1,3 @@
-import validate from "."
-
 Object.entries = function (obj) { //eslint-disable-line no-extend-native
   let ownProps = Object.keys(obj),
     i = ownProps.length,
@@ -27,44 +25,3 @@ export function dualogParse(string) {
       .map(e => e.split("/"))
   )
 }
-
-/**
- *
- * @param {object} message
- */
-export function validateMessage(message){
-  console.log(`Validating DEP message ${message}`)
-
-  const requiredFields = ["TM", "RN", "RC", "MA", "DA", "TI", "PO", "ZD", "ZT", "OB", "PD", "PT", "LA", "LO", "AC", "DS"]
-
-  let result = {}
-
-  switch (message.TM) {
-  case "DEP":
-    // For hvert element i required
-    requiredFields.forEach(field => {
-      if(!message[field]){
-        console.log(`${field} not found`)
-        result = {...result, RS: "NAK", RE: 104}
-      } else {
-        const fieldResult = validate[field](message[field])
-        if (fieldResult) {
-          result = {...result, ...fieldResult, RS: "NAK"}
-        }
-      }
-    })
-
-    break
-
-  default:
-    result = {...result, RE: 530, RS: "NAK"}
-    break
-  }
-  console.log("Field checking is done")
-
-  if (result.RS === "NAK") return result
-  else return {RS: "ACK"}
-
-}
-
-
